requirements.txt
robotframework-seleniumlibrary==6.5.0
robotframework-tidy